pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = "A to-do list and task manager for professionals and small businesses. Combining tasks, projects, comments, attachments, notifications, and more."
authors = ["Artem <artem.mikhaylov5593@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.2"
djangorestframework-simplejwt = "^5.2.2"
psycopg2 = "^2.9.5"
django-environ = "^0.9.0"
social-auth-app-django = "^5.1.0"
django-filter = "^23.1"
pydantic = "^1.10.7"
pytest-factoryboy = "^2.5.1"
pytest = "^7.3.1"
pytest-django = "^4.5.2"
pytest-faker = "^2.0.0"
pytest-random-order = "^1.1.0"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
ansible-vault-win = "^1.0.0"
django-extensions = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
